+++ Cargo.toml
@@ -18,8 +18,6 @@ parking_lot = { workspace = true }
 dashmap = { workspace = true }
 
 tracing-subscriber = "0.3.18"
-
-tokio-tungstenite = { version = "0.21", features = ["rustls-tls-webpki-roots"], optional = true  }
 ocrs.workspace = true
 rten.workspace = true
 
@@ -63,5 +61,5 @@ branch = "serenity-next-merged"
 [profile.release]
 strip = true
 opt-level = 3
-#lto = true
-#codegen-units = 1
+lto = true
+codegen-units = 1
